Update allowing remote management for Jira issues
The patch provides a lot of extended features for Jira Issue management.
None of the current functionality has been removed which means the
plug-in should be backward compatible. There are more features that will
need to be added to fully implement the API however this feature set
will allow most developers to access the normal level of access
necessary for a remote API interface for Jira.
Highlights:
1) Issues can be created, updated, or deleted as needed. This includes
most components such as comments, transitions, and assignees.
2) Projects can now be accessed as well as related issues
3) A Field class has been added to allow proper access to the meta data
for create, edit, and update requests. This data is searchable via Hash
or dot notation
4) Some sections can now be lazy loaded to a degree
a. Project requests will not load tickets just the ID’s and Keys. This
reduces the download every time the project is called.
i. When the issue is accessed and is not fully loaded it will load the
issue and attempt the request again
b. Issues Lazy loading
i. Lazy loading has been implemented for meta data
ii. Available for other aspects such as watchers and comments
c. Lazy loading requires the overriding of the “reload” function in
Jiralicious::Base without this it is not enabled.
5) Jiralicious::Field class
a. This is a new class that is used to manage the meta data however it
could be used in other places
b. The class is designed to parse and load any Hash/Array object
c. WARNING: The inspect on the root level of this object can be quite
extensive and can be difficult to read.

Jira REST API features added or updated
/customFieldOption/{id} [GET]
/field [GET]
/issue [POST]
/issue/createmeta [GET]
/issue/{issueIdOrKey} [GET, DELETE, PUT]
/issue/{issueIdOrKey}/assignee [PUT]
/issue/{issueIdOrKey}/comment [GET, POST]
/issue/{issueIdOrKey}/comment/{id} [GET, PUT, DELETE]
/issue/{issueIdOrKey}/editmeta [GET]
/issue/{issueIdOrKey}/transitions [GET, POST]
/issue/{issueIdOrKey}/watchers [GET, POST, DELETE]
/project [GET]
/project/{key} [GET]