Update bundler in travis before build
Tests are failing in Ruby <1.9.3 due to Travis' version of Bundler being
outdated. We ran into this issue with one of our repositories the other
day and were able to solve it by adding:

before_install:
  - gem update bundler

See bundler/bundler#3558 for more information

I made the change to this PR to get the build to pass. Let me know if
you would like me to do that separately. These two changes should
unbreak master.